docs: curate the agent routing map and separate the full index
